+++ libs/audio/targets/snd_oss.c
@@ -186,7 +186,7 @@ SNDDMA_Init (void)
 	if (snd_rate->int_val) {
 		shm->speed = snd_rate->int_val;
 	} else {
-		for (i = 0; i < (sizeof (tryrates) / 4); i++)
+		for (i = 0; i < ((int) sizeof (tryrates) / 4); i++)
 			if (!ioctl (audio_fd, SNDCTL_DSP_SPEED, &tryrates[i]))
 				break;
 		shm->speed = tryrates[i];
