+++ libs/models/brush/model_brush.c
@@ -819,7 +819,7 @@ Mod_LoadBrushModel (model_t *mod, void *buffer)
 	// swap all the lumps
 	mod_base = (byte *) header;
 
-	for (i = 0; i < sizeof (dheader_t) / 4; i++)
+	for (i = 0; i < (int) sizeof (dheader_t) / 4; i++)
 		((int *) header)[i] = LittleLong (((int *) header)[i]);
 
 	// checksum all of the map, except for entities
