+++ libs/video/renderer/sw/screen.c
@@ -177,8 +177,10 @@ SCR_ScreenShot (int width, int height)
 	// enable direct drawing of console to back buffer
 	D_EnableBackBufferAccess ();
 
-	w = (vid.width < width) ? vid.width : width;
-	h = (vid.height < height) ? vid.height : height;
+	w = (vid.width < (unsigned int) width) ? vid.width
+										   : (unsigned int) width;
+	h = (vid.height < (unsigned int) height) ? vid.height
+											 : (unsigned int) height;
 
 	fracw = (float) vid.width / (float) w;
 	frach = (float) vid.height / (float) h;
