+++ libs/video/renderer/sw/sw_rdraw.c
@@ -400,7 +400,7 @@ R_RenderFace (msurface_t *fa, int clipflags)
 			if (!insubmodel) {
 				if (r_pedge->cachededgeoffset & FULLY_CLIPPED_CACHED) {
 					if ((r_pedge->cachededgeoffset & FRAMECOUNT_MASK) ==
-						r_framecount) {
+						(unsigned int) r_framecount) {
 						r_lastvertvalid = false;
 						continue;
 					}
@@ -435,7 +435,7 @@ R_RenderFace (msurface_t *fa, int clipflags)
 			if (!insubmodel) {
 				if (r_pedge->cachededgeoffset & FULLY_CLIPPED_CACHED) {
 					if ((r_pedge->cachededgeoffset & FRAMECOUNT_MASK) ==
-						r_framecount) {
+						(unsigned int) r_framecount) {
 						r_lastvertvalid = false;
 						continue;
 					}
