+++ libs/video/renderer/sw32/sw32_rdraw.c
@@ -393,7 +393,7 @@ R_RenderFace (msurface_t *fa, int clipflags)
 			if (!insubmodel) {
 				if (r_pedge->cachededgeoffset & FULLY_CLIPPED_CACHED) {
 					if ((r_pedge->cachededgeoffset & FRAMECOUNT_MASK) ==
-						r_framecount) {
+						(unsigned int) r_framecount) {
 						r_lastvertvalid = false;
 						continue;
 					}
@@ -428,7 +428,7 @@ R_RenderFace (msurface_t *fa, int clipflags)
 			if (!insubmodel) {
 				if (r_pedge->cachededgeoffset & FULLY_CLIPPED_CACHED) {
 					if ((r_pedge->cachededgeoffset & FRAMECOUNT_MASK) ==
-						r_framecount) {
+						(unsigned int) r_framecount) {
 						r_lastvertvalid = false;
 						continue;
 					}
