+++ libs/video/targets/vid_3dfxsvga.c
@@ -253,7 +253,7 @@ static int  resolutions[][3] = {
 static int
 findres (int *width, int *height)
 {
-	int         i;
+	unsigned int i;
 
 	for (i = 0; i < NUM_RESOLUTIONS; i++) {
 		if ((*width <= resolutions[i][0]) && (*height <= resolutions[i][1])) {
@@ -322,8 +322,9 @@ VID_Init (unsigned char *palette)
 
 	qf_fxMesaMakeCurrent (fc);
 
-	vid.width = vid.conwidth = min (vid.conwidth, scr_width);
-	vid.height = vid.conheight = min (vid.conheight, scr_height);
+	vid.width = vid.conwidth = min (vid.conwidth, (unsigned int) scr_width);
+	vid.height = vid.conheight = min (vid.conheight,
+									  (unsigned int) scr_height);
 
 	vid.aspect = ((float) vid.height / (float) vid.width) * (320.0 / 240.0);
 	vid.numpages = 2;
